Refactor structure repo
