+++ src/main/java/com/example/demo/DemoBookApplication.java
@@ -1,4 +1,4 @@
-package com.example.demoBook;
+package com.example.demo;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
