+++ src/main/java/com/example/demo/configuration/Authentication.java
@@ -1,4 +1,4 @@
-package com.example.demoBook.configuration;
+package com.example.demo.configuration;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
