+++ src/main/java/com/example/demo/configuration/JooqConfigMetaData.java
@@ -1,4 +1,4 @@
-package com.example.demoBook.configuration;
+package com.example.demo.configuration;
 
 import java.util.Properties;
 import javax.sql.DataSource;
