+++ src/main/java/com/example/demo/controller/BookController.java
@@ -1,4 +1,4 @@
-package com.example.demoBook.controller;
+package com.example.demo.controller;
 
 import java.util.List;
 
@@ -6,8 +6,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 
 import org.springframework.web.bind.annotation.*;
 
-import com.example.demoBook.book.Book;
-import com.example.demoBook.service.BookServiceJooq;
+import com.example.demo.model.Book;
+import com.example.demo.service.BookServiceJooq;
 
 @RestController
 @RequestMapping(path = "/api/v1/books")
