+++ src/main/java/com/example/demo/controller/BookControllerV2.java
@@ -1,12 +1,12 @@
-package com.example.demoBook.controller;
+package com.example.demo.controller;
 
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
-import com.example.demoBook.book.Book;
-import com.example.demoBook.service.BookServiceJooqV2;
+import com.example.demo.model.Book;
+import com.example.demo.service.BookServiceJooqV2;
 
 @RestController
 @RequestMapping(path = "/api/v2/books")
