+++ src/main/java/com/example/demo/exception/ExceptionData.java
@@ -1,6 +1,5 @@
-package com.example.demoBook.exceptions;
+package com.example.demo.exception;
 
-import lombok.AllArgsConstructor;
 import lombok.Data;
 
 @Data
