+++ src/main/java/com/example/demo/model/Book.java
@@ -1,4 +1,4 @@
-package com.example.demoBook.book;
+package com.example.demo.model;
 
 import java.time.LocalDate;
