+++ src/test/java/com/example/demo/DemoBookApplicationTests.java
@@ -1,4 +1,4 @@
-package com.example.demoBook;
+package com.example.demo;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
